quad

+++ inc/Engine/IndexBuffer.hpp
@@ -20,7 +20,7 @@ class IndexBuffer
 	IndexBuffer& operator=(IndexBuffer&& other) noexcept;
 
 	void bind() const;
-	void loadData(size_t length, const GLfloat* pointer, GLenum usage) const;
-	void loadData(std::span<GLfloat> data, GLenum usage) const;
+	void loadData(size_t length, const GLuint* pointer, GLenum usage) const;
+	void loadData(std::span<GLuint> data, GLenum usage) const;
 };
 } // namespace phyto::engine
