+++ src/Engine/IndexBuffer.cpp
@@ -39,13 +39,13 @@ void phyto::engine::IndexBuffer::bind() const
 	glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, m_id);
 }
 
-void phyto::engine::IndexBuffer::loadData(size_t length, const GLfloat* pointer, GLenum usage) const
+void phyto::engine::IndexBuffer::loadData(size_t length, const GLuint* pointer, GLenum usage) const
 {
 	bind();
 	glBufferData(GL_ELEMENT_ARRAY_BUFFER, static_cast<GLsizeiptr>(length), pointer, usage);
 }
 
-void phyto::engine::IndexBuffer::loadData(std::span<GLfloat> data, GLenum usage) const
+void phyto::engine::IndexBuffer::loadData(std::span<GLuint> data, GLenum usage) const
 {
 	loadData(data.size_bytes(), data.data(), usage);
 }
